Extend row focusing pattern to image metadata
