+++ arches/app/src/components/ControlledListManager/editor/ImageEditor.vue
@@ -1,21 +1,12 @@
 <script setup lang="ts">
 import arches from "arches";
 import { computed, inject } from "vue";
-import { useGettext } from "vue3-gettext";
 
-import Button from "primevue/button";
-import { useToast } from "primevue/usetoast";
-
-import { deleteImage } from "@/components/ControlledListManager/api.ts";
 import {
-    DANGER,
-    DEFAULT_ERROR_TOAST_LIFE,
-    ERROR,
-    METADATA_CHOICES,
     itemKey,
+    METADATA_CHOICES,
 } from "@/components/ControlledListManager/constants.ts";
 import { bestLabel } from "@/components/ControlledListManager/utils.ts";
-import AddMetadata from "@/components/ControlledListManager/editor/AddMetadata.vue";
 import ImageMetadata from "@/components/ControlledListManager/editor/ImageMetadata.vue";
 
 import type { Ref } from "vue";
@@ -26,28 +17,6 @@ import type {
 
 const item = inject(itemKey) as Ref<ControlledListItem>;
 const { image } = defineProps<{ image: ControlledListItemImage }>();
-
-const toast = useToast();
-const { $gettext } = useGettext();
-
-const labeledChoices = [
-    {
-        type: METADATA_CHOICES.title,
-        label: $gettext("Title"),
-    },
-    {
-        type: METADATA_CHOICES.alternativeText,
-        label: $gettext("Alternative text"),
-    },
-    {
-        type: METADATA_CHOICES.description,
-        label: $gettext("Description"),
-    },
-    {
-        type: METADATA_CHOICES.attribution,
-        label: $gettext("Attribution"),
-    },
-];
 
 const bestTitle = computed(() => {
     const titles = image.metadata.filter(
@@ -71,27 +40,6 @@ const bestAlternativeText = computed(() => {
             ?.value || bestLabel(item.value, arches.activeLanguage).value
     );
 });
-
-const issueDeleteImage = async () => {
-    try {
-        await deleteImage(image);
-        removeImage(image);
-    } catch (error) {
-        toast.add({
-            severity: ERROR,
-            life: DEFAULT_ERROR_TOAST_LIFE,
-            summary: $gettext("Image deletion failed"),
-            detail: error instanceof Error ? error.message : undefined,
-        });
-    }
-};
-
-const removeImage = (removedImage: ControlledListItemImage) => {
-    const toDelete = item.value!.images.findIndex(
-        (imageFromItem) => imageFromItem.id === removedImage.id,
-    );
-    item.value!.images.splice(toDelete, 1);
-};
 </script>
 
 <template>
@@ -102,37 +50,6 @@ const removeImage = (removedImage: ControlledListItemImage) => {
             :alt="bestAlternativeText"
             width="200"
         />
-        <div>
-            <ImageMetadata
-                :metadata="image.metadata"
-                :labeled-choices
-            />
-            <div style="display: flex; gap: 1rem">
-                <AddMetadata
-                    :image
-                    :labeled-choices
-                />
-                <Button
-                    raised
-                    :severity="DANGER"
-                    icon="fa fa-trash"
-                    :label="$gettext('Delete image')"
-                    @click="issueDeleteImage"
-                />
-            </div>
-        </div>
+        <ImageMetadata :image="image" />
     </div>
 </template>
-
-<style scoped>
-.p-button {
-    height: 3rem;
-    margin-top: 1rem;
-}
-:deep(.p-button-icon),
-:deep(.p-button-label) {
-    color: white;
-    font-size: small;
-    font-weight: 600;
-}
-</style>
